add NULL protection and implememted more functions

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: mkeerewe <mkeerewe@student.42lausanne.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/19 16:43:08 by mkeerewe          #+#    #+#             */
-/*   Updated: 2025/08/19 16:43:09 by mkeerewe         ###   ########.fr       */
+/*   Updated: 2025/08/21 17:38:40 by mkeerewe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int		i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
